Hola mundo 3

+++ src/main/java/org/jrae/kinal_play/dominio/dto/ModPeliculaDto.java
@@ -1,10 +1,11 @@
 package org.jrae.kinal_play.dominio.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record ModPeliculaDto(
         String name,
         LocalDate releaseDate,
-        Integer rating
+        Double rating
 ) {
 }
